Add files via upload
This code contains 10 different wishes for an Indian festival. When the user says "Happy Sankranthi India" to an amazon echo device or to an app on a smartphone, the skill will be invoked and random wish is chosen from among the list of 10 different wishes and given back to the user as the output.